Added Student Grades
Added support for each student's name to have a
grade added to it. Constructor updated for this.
Test cases updated to check for required fields
of this support.
Completed feature to remove student's name from
the list.
Modified buttons and labels in fxml file to read
more easily and formatted per OCD requirements.