Implemented Recipe conversion/added test cases.
Created static method convertRecipeToString() to
FileWriter class. Converts a recipe to a string
with the specified format. Name should appear on
first line and a CSV ingredients list of names on
the second.
Throws IAE if either recipe/ingredients is null.
Added test cases for convertRecipeToString() and
writeRecipeToFile() checking for proper function
of formatting recipe name/ingredients and for the
successful writing of recipe to a file. Includes
testing for proper exception handling of null or
empty scenarios.